requirements.txt
Appium-Python-Client>=3.0.0
Selenium>=4.12.0
pytest>=7.4.2
allure-pytest>=2.13.2
selene>=2.0.0rc4
pydantic~=2.4.0